Rename buttons to feature
