+++ plugins/talk-plugin-rich-text/client/components/rte/features/Blockquote.js
