+++ plugins/talk-plugin-rich-text/client/components/rte/features/Italic.js
